requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
WTForms-SQLAlchemy==0.3

# DB
PyMySQL==1.1.0
psycopg2-binary>=2.9.9,<2.10

# Utilità
cryptography>=42.0.0,<45
python-dotenv==1.0.0
email-validator==2.0.0
imapclient==2.3.1

# Scheduler
APScheduler==3.10.4
